Move and rename MasterInfoIntegrationTest
